requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2023.11.17
click==8.1.7
distro==1.9.0
Flask==3.0.0
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
openai==1.7.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
